apps/processor/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "opencv-python-headless>=4.13.0.92",
    "python-dotenv>=1.2.2",
    "openai-agents>=0.5.4",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "Pillow>=10.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "python-multipart>=0.0.9",
]
